create function for pictures templates
